Put submitted proposal tests in own module.
